feat: manage blocked user public chat and online status
